+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/sharing/ShareData.kt
@@ -61,7 +61,7 @@ import de.connect2x.trixnity.messenger.util.FileDescriptor
 import de.connect2x.trixnity.messenger.util.SharedData
 import de.connect2x.trixnity.messenger.viewmodel.sharing.ShareDataViewModel
 import de.connect2x.trixnity.messenger.viewmodel.util.formatSize
-import de.connect2x.trixnity.messenger.viewmodel.util.limitedByteArrayOrNull
+import net.folivo.trixnity.utils.toByteArray
 
 interface ShareDataView {
     @Composable
@@ -209,7 +209,7 @@ private fun ShareUrlRow(text: String, icon: FileDescriptor?, maxMediaSize: Long)
     var image by remember { mutableStateOf<ImageBitmap?>(null) }
 
     LaunchedEffect(icon) {
-        icon?.content?.limitedByteArrayOrNull(maxMediaSize)
+        icon?.content?.toByteArray(maxMediaSize)
             ?.also { image = it.toImageBitmap() }
     }