+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/CreateMediaStoreModule.kt
@@ -1,0 +1,10 @@
+package de.connect2x.trixnity.messenger
+
+import net.folivo.trixnity.core.model.UserId
+import org.koin.core.module.Module
+
+fun interface CreateMediaStoreModule {
+    suspend operator fun invoke(userId: UserId): Module
+}
+
+expect fun platformCreateMediaStoreModuleModule(): Module