+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/room/timeline/elements/message/ImageRoomMessageTimelineElementViewModel.kt
@@ -42,15 +42,15 @@ class ImageRoomMessageTimelineElementViewModelImpl(
 ) : RoomMessageTimelineElementViewModel.FileBased.Image,
     FileBasedRoomMessageTimelineElementViewModel<FileBased.Image>(viewModelContext, content) {
 
-    private val maxPreviewSize = get<MatrixMessengerConfiguration>().maxMediaSizeInMemory
     private val thumbnails = get<Thumbnails>()
 
     private val thumbnailProgressFlow = MutableStateFlow<FileTransferProgress?>(null)
+    private val maxMediaSizeInMemory = get<MatrixMessengerConfiguration>().maxMediaSizeInMemory
 
     override val thumbnail: StateFlow<ByteArray?> = flow {
         emit(
             // TODO needs some sort of retry!
-            thumbnails.loadThumbnail(matrixClient, content, thumbnailProgressFlow, maxPreviewSize)
+            thumbnails.loadThumbnail(coroutineScope, matrixClient, content, thumbnailProgressFlow, maxMediaSizeInMemory)
         )
     }.stateIn(coroutineScope, whileSubscribedWithTimeout, null)