+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/settings/AvatarCutterViewModel.kt
@@ -5,7 +5,6 @@ import de.connect2x.trixnity.messenger.MatrixMessengerConfiguration
 import de.connect2x.trixnity.messenger.i18n.I18n
 import de.connect2x.trixnity.messenger.util.FileDescriptor
 import de.connect2x.trixnity.messenger.viewmodel.MatrixClientViewModelContext
-import de.connect2x.trixnity.messenger.viewmodel.util.limitedByteArrayOrNull
 import io.github.oshai.kotlinlogging.KotlinLogging
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -14,6 +13,7 @@ import kotlinx.coroutines.launch
 import net.folivo.trixnity.client.media
 import net.folivo.trixnity.core.model.RoomId
 import net.folivo.trixnity.core.model.events.m.room.AvatarEventContent
+import net.folivo.trixnity.utils.toByteArray
 import org.koin.core.component.get
 
 
@@ -52,7 +52,6 @@ open class AvatarCutterViewModelImpl(
 ) : MatrixClientViewModelContext by viewModelContext, AvatarCutterViewModel {
 
     private val i18n = get<I18n>()
-    private val messengerConfiguration = get<MatrixMessengerConfiguration>()
 
     override val upload = MutableStateFlow(false)
     override val error = MutableStateFlow<String?>(null)
@@ -65,19 +64,18 @@ open class AvatarCutterViewModelImpl(
         cancel()
     }
 
-    override val maxAvatarSize: Long = messengerConfiguration.avatarMaxSize
+    private val maxMediaSizeInMemory = get<MatrixMessengerConfiguration>().maxMediaSizeInMemory
+    override val maxAvatarSize: Long = maxMediaSizeInMemory
 
     private val _avatarImage: MutableStateFlow<ByteArray?> = MutableStateFlow(null)
+
 
     init {
         backHandler.register(backCallback)
         coroutineScope.launch {
             val fileSize = file.fileSize
             if (fileSize == null || fileSize <= maxAvatarSize) {
-                _avatarImage.value = file.content.limitedByteArrayOrNull(maxAvatarSize) {
-                    log.warn { "Uploaded avatar file exceeds avatar size limits, so it's not shown" }
-                    error.value = i18n.avatarSizeMaxSizeError(maxAvatarSize)
-                }
+                _avatarImage.value = file.content.toByteArray(maxMediaSizeInMemory)
             } else {
                 log.warn { "Uploaded avatar file exceeds avatar size limits, so it's not shown" }
                 error.value = i18n.avatarSizeMaxSizeError(maxAvatarSize)