+++ trixnity-messenger/src/jsMain/kotlin/de/connect2x/trixnity/messenger/CreateMediaStore.js.kt
@@ -2,8 +2,8 @@ package de.connect2x.trixnity.messenger
 
 import de.connect2x.trixnity.messenger.util.RootPath
 import io.github.oshai.kotlinlogging.KotlinLogging
-import net.folivo.trixnity.client.media.indexeddb.IndexedDBMediaStore
-import net.folivo.trixnity.client.media.opfs.OpfsMediaStore
+import net.folivo.trixnity.client.media.indexeddb.createIndexedDBMediaStoreModule
+import net.folivo.trixnity.client.media.opfs.createOpfsMediaStoreModule
 import org.koin.core.module.Module
 import org.koin.dsl.module
 import web.fs.FileSystemGetDirectoryOptions
@@ -11,10 +11,10 @@ import web.navigator.navigator
 
 private val log = KotlinLogging.logger {}
 
-actual fun platformCreateMediaStoreModule(): Module = module {
-    single<CreateMediaStore> {
+actual fun platformCreateMediaStoreModuleModule(): Module = module {
+    single<CreateMediaStoreModule> {
         val rootPath = get<RootPath>()
-        CreateMediaStore { userId ->
+        CreateMediaStoreModule { userId ->
             val basePath = rootPath.forAccountMedia(userId)
             try {
                 var opfsDirectory = navigator.storage.getDirectory()
@@ -22,10 +22,10 @@ actual fun platformCreateMediaStoreModule(): Module = module {
                     opfsDirectory =
                         opfsDirectory.getDirectoryHandle(segment, FileSystemGetDirectoryOptions(create = true))
                 }
-                OpfsMediaStore(opfsDirectory)
+                createOpfsMediaStoreModule(opfsDirectory)
             } catch (error: Throwable) {
                 log.warn(error) { "failed to use OPFS as MediaStore. Falling back to Indexeddb" }
-                IndexedDBMediaStore(basePath.toString())
+                createIndexedDBMediaStoreModule(basePath.toString())
             }
         }
     }